requirements.txt
aiohttp==3.9.5
aiogram==3.21.0
psutil==5.9.0
feedparser==6.0.10
Pillow==10.2.0
python-dotenv==1.0.0
validators==0.22.0
python-Levenshtein==0.25.0
defusedxml==0.7.1
python-dateutil==2.9.0.post0
yarl==1.9.4
colorama==0.4.6
beautifulsoup4==4.12.2
soupsieve==2.0.1
aiofiles==24.1.0
typing-extensions==4.14.1
pydantic==2.11.7
matplotlib==3.8.0
numpy==1.26.0
requests==2.31.0
python-telegram-bot==20.3
pytz==2025.2